requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
altgraph==0.17.3
annotated-types==0.6.0
anyio==3.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.6.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
audioread==3.0.1
auto-py-to-exe==2.25.0
Babel==2.14.0
beautifulsoup4==4.12.2
bert-extractive-summarizer==0.10.1
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
boto3==1.34.34
botocore==1.34.34
bottle==0.12.23
bottle-websocket==0.2.9
breadability==0.1.20
cachetools==5.3.2
catalogue==2.0.10
certifi==2024.7.4
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.3
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
confection==0.1.4
contextlib2==21.6.0
contourpy==1.2.0
cryptography==41.0.1
customtkinter==5.2.0
cvlib==0.2.7
cycler==0.12.1
cymem==2.0.8
darkdetect==0.8.0
dask==2024.9.0
dask-expr==1.1.14
dask-glm==0.3.2
dask-ml==2024.4.4
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distributed==2024.9.0
distro==1.9.0
Django==4.1.7
django-admin-rangefilter==0.9.0
docopt==0.6.2
Eel==0.14.0
execnet==1.9.0
executing==2.0.1
fastapi==0.98.0
fastjsonschema==2.19.1
filelock==3.9.0
Flask==1.1.4
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.2
gast==0.5.4
gensim==4.3.2
gevent==22.10.2
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googletrans==3.0.0
graphviz==0.20.3
greenlet==2.0.1
grpcio==1.60.1
h11==0.14.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2022.11.1
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
hyperframe==5.2.0
idna==3.4
imageio==2.24.0
importlib-metadata==7.0.1
imutils==0.5.4
iniconfig==1.1.1
ipykernel==6.29.0
ipython==8.21.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==1.1.0
jedi==0.19.1
Jinja2==2.11.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kaggle==1.6.17
keras==2.15.0
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.4
libclang==16.0.6
libretranslatepy==2.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
locket==1.0.0
loguru==0.6.0
lxml==5.0.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mock==4.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multipledispatch==1.0.0
murmurhash==1.0.10
mysql-connector-python==9.1.0
Naked==0.1.32
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.2.1
nltk==3.8.1
notebook==7.0.7
notebook_shim==0.2.3
notify-py==0.3.42
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
opencv-contrib-python==4.6.0.66
opencv-python==4.7.0.68
opt-einsum==3.3.0
overrides==7.7.0
packaging==21.3
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
partd==1.4.2
path==16.5.0
path.py==12.5.0
patsy==0.5.6
pdfkit==1.0.0
pefile==2023.2.7
Pillow==9.3.0
pipdeptree==2.9.3
platformdirs==2.6.2
plotly==5.20.0
pluggy==1.0.0
plyer==2.1.0
pooch==1.8.2
preshed==3.0.9
progressbar==2.5
projectq==0.8.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.21.12
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycountry==23.12.11
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.9
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.4
PyMySQL==1.1.1
pyparsing==3.0.9
PyPDF2==2.11.2
Pyqtree==1.0.0
pytest==7.2.0
pytest-shutil==1.7.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2022.6
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.12
PyYAML==6.0
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
Send2Trash==1.8.2
shellescape==3.8.1
six==1.16.0
sklearn-model==0.0.6
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.5.0.post1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sparse==0.15.4
sqlparse==0.4.3
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
statsmodels==0.14.1
streamlit==1.39.0
sumy==0.11.0
sympy==1.12
tblib==3.0.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.1.1
terminado==0.18.0
text-unidecode==1.3
thinc==8.2.2
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
tzdata==2022.7
tzlocal==5.2
umap==0.1.1
uri-template==1.3.0
urllib3==2.0.3
uvicorn==0.22.0
validators==0.22.0
virtualenv==20.17.1
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==1.0.1
whichcraft==0.6.1
widgetsnbextension==4.0.9
win32-setctime==1.1.0
wrapt==1.14.1
yarl==1.9.4
zict==3.0.0
zipp==3.17.0
zope.event==4.5.0
zope.interface==5.5.2
